Jsonifier Release v0.9.7
-Removed a bunch of superfluous "inline" calls.
-Updated a few types.
-Updated the Vector class to fill its implementation out.
-Updated the Iterator and ConstIterator implementations as well as the iterator-related members of the String, StringView and Vector classes.

+++ Include/jsonifier/RawArray.hpp
@@ -1,22 +1,22 @@
 /*
-		MIT License
+	MIT License
 
 	Copyright (c) 2023 RealTimeChris
 
-	Permission is hereby granted, free of charge, to any person obtaining a copy of this 
-	software and associated documentation files (the "Software"), to deal in the Software 
-	without restriction, including without limitation the rights to use, copy, modify, merge, 
-	publish, distribute, sublicense, and/or sell copies of the Software, and to permit 
+	Permission is hereby granted, free of charge, to any person obtaining a copy of this
+	software and associated documentation files (the "Software"), to deal in the Software
+	without restriction, including without limitation the rights to use, copy, modify, merge,
+	publish, distribute, sublicense, and/or sell copies of the Software, and to permit
 	persons to whom the Software is furnished to do so, subject to the following conditions:
 
-	The above copyright notice and this permission notice shall be included in all copies or 
+	The above copyright notice and this permission notice shall be included in all copies or
 	substantial portions of the Software.
 
-	THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, 
-	INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR 
-	PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE 
-	FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR 
-	OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER 
+	THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+	INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR
+	PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE
+	FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
+	OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 	DEALINGS IN THE SOFTWARE.
 */
 /// https://github.com/RealTimeChris/Jsonifier
@@ -29,28 +29,28 @@ namespace JsonifierInternal {
 
 	template<typename ValueType, size_t N> struct RawArray {
 	  public:
-		using value_type = ValueType;
-		using reference = value_type&;
+		using value_type	  = ValueType;
+		using reference		  = value_type&;
 		using const_reference = const value_type&;
-		using pointer = value_type*;
-		using const_pointer = const value_type*;
-		using iterator = pointer;
-		using const_iterator = const_pointer;
-		using size_type = size_t;
+		using pointer		  = value_type*;
+		using const_pointer	  = const value_type*;
+		using iterator		  = pointer;
+		using const_iterator  = const_pointer;
+		using size_type		  = uint64_t;
 		using difference_type = std::ptrdiff_t;
 
-		inline constexpr RawArray() noexcept = default;
-		inline constexpr RawArray& operator=(RawArray&&) noexcept = default;
-		inline constexpr RawArray(RawArray&&) noexcept = default;
+		inline constexpr RawArray() noexcept						   = default;
+		inline constexpr RawArray& operator=(RawArray&&) noexcept	   = default;
+		inline constexpr RawArray(RawArray&&) noexcept				   = default;
 		inline constexpr RawArray& operator=(const RawArray&) noexcept = default;
-		inline constexpr RawArray(const RawArray&) noexcept = default;
+		inline constexpr RawArray(const RawArray&) noexcept			   = default;
 
 		template<size_t M> inline constexpr RawArray(ValueType const (&init)[M]) : RawArray(init, std::make_index_sequence<N>()) {
 			static_assert(M >= N);
 		}
 
 		inline constexpr RawArray(const std::initializer_list<ValueType>& other) {
-			for (size_t x = 0; x < other.size(); ++x) {
+			for (uint64_t x = 0; x < other.size(); ++x) {
 				operator[](x) = std::move(other.begin()[x]);
 			}
 		}
@@ -79,22 +79,22 @@ namespace JsonifierInternal {
 			return N;
 		}
 
-		inline constexpr reference operator[](size_t index) {
+		inline constexpr reference operator[](uint64_t index) {
 			return dataVal[index];
 		}
 
-		inline constexpr const_reference operator[](size_t index) const noexcept {
+		inline constexpr const_reference operator[](uint64_t index) const noexcept {
 			return dataVal[index];
 		}
 
-		inline constexpr reference at(size_t index) noexcept {
+		inline constexpr reference at(uint64_t index) noexcept {
 			if (index > N) {
 				std::abort();
 			}
 			return dataVal[index];
 		}
 
-		inline constexpr const_reference at(size_t index) const noexcept {
+		inline constexpr const_reference at(uint64_t index) const noexcept {
 			if (index > N) {
 				std::abort();
 			}
@@ -126,7 +126,7 @@ namespace JsonifierInternal {
 		}
 
 		inline constexpr void fill(const value_type& val) {
-			for (size_t x = 0; x < N; ++x) {
+			for (uint64_t x = 0; x < N; ++x) {
 				dataVal[x] = val;
 			}
 		}
@@ -138,17 +138,17 @@ namespace JsonifierInternal {
 
 	template<typename ValueType> class RawArray<ValueType, 0> {
 	  public:
-		using value_type = ValueType;
-		using reference = value_type&;
-		using const_reference = const value_type&;
-		using pointer = value_type*;
-		using const_pointer = const value_type*;
-		using iterator = pointer;
-		using const_iterator = const_pointer;
-		using reverse_iterator = std::reverse_iterator<iterator>;
+		using value_type			 = ValueType;
+		using reference				 = value_type&;
+		using const_reference		 = const value_type&;
+		using pointer				 = value_type*;
+		using const_pointer			 = const value_type*;
+		using iterator				 = pointer;
+		using const_iterator		 = const_pointer;
+		using reverse_iterator		 = std::reverse_iterator<iterator>;
 		using const_reverse_iterator = std::reverse_iterator<const_iterator>;
-		using size_type = size_t;
-		using difference_type = std::ptrdiff_t;
+		using size_type				 = uint64_t;
+		using difference_type		 = std::ptrdiff_t;
 
 		inline constexpr RawArray() noexcept = default;
 	};
